+++ CHANGELOG.md
@@ -4,14 +4,13 @@
 - Fix project CI: linter, tests, build, release
 - Add pass-through options usage hint
 - Return correct exit code when wrong CLI params supplied
-- Upgrade Go version to 1.17.8
 - Cleanup Travis config and make it use the list of actual ArangoD versions
 - Add documentation header to generated command files
 - Add support for ArangoD exit codes. Do not try to restart instance if exit code is unrecoverable
 - Show process output in case there was an error during process startup
 - Deprecate --feature.jwt.rotation CLI option
 - Add support for config file
-- Switch to Go 1.17.13 and alpine3.16 for build image
+- Upgrade to Go 1.17.13 and alpine3.16 for build image
 
 ## [0.15.4](https://github.com/arangodb-helper/arangodb/tree/0.15.4) (2022-03-22)
 - Use github.com/golang-jwt/jwt